IN-337: refactor: some filtering logic has been moved to the hook useImageFilter; separate slide component has been created; feat: images have been added to the filter buttons
