+++ package.json
@@ -23,7 +23,6 @@
     "autoprefixer": "^10.4.20",
     "colorthief": "^2.6.0",
     "date-fns": "^4.1.0",
-    "dom-to-image": "^2.6.0",
     "framer-motion": "^11.5.4",
     "jose": "^5.9.4",
     "lodash": "^4.17.21",
