+++ src/features/create-post/utils/filterValues.ts
@@ -1,175 +1,47 @@
-export type FilterValue = {
-  class: string
-  name: string
+import { Filters } from '../types'
+
+export type FilterValues = {
+  name: Filters
 }
 
-export const filterValues: FilterValue[] = [
+export const filterValues: FilterValues[] = [
   {
-    class: '',
-    name: 'None',
+    name: 'original',
   },
   {
-    class: 'filter-1977',
-    name: '1977',
+    name: 'posterize',
   },
   {
-    class: 'filter-aden',
-    name: 'Aden',
+    name: 'vignette',
   },
   {
-    class: 'filter-amaro',
-    name: 'Amaro',
+    name: 'heatmap',
   },
   {
-    class: 'filter-ashby',
-    name: 'Ashby',
+    name: 'contrast',
   },
   {
-    class: 'filter-brannan',
-    name: 'Brannan',
+    name: 'brightness',
   },
   {
-    class: 'filter-brooklyn',
-    name: 'Brooklyn',
+    name: 'pixelate',
   },
   {
-    class: 'filter-charmes',
-    name: 'Charmes',
+    name: 'sepia',
   },
   {
-    class: 'filter-clarendon',
-    name: 'Clarendon',
+    name: 'grayscale',
   },
   {
-    class: 'filter-crema',
-    name: 'Crema',
+    name: 'negative',
   },
   {
-    class: 'filter-dogpatch',
-    name: 'Dogpatch',
+    name: 'threshold',
   },
   {
-    class: 'filter-earlybird',
-    name: 'Earlybird',
+    name: 'sharpen',
   },
   {
-    class: 'filter-gingham',
-    name: 'Gingham',
-  },
-  {
-    class: 'filter-ginza',
-    name: 'Ginza',
-  },
-  {
-    class: 'filter-hefe',
-    name: 'Hefe',
-  },
-  {
-    class: 'filter-helena',
-    name: 'Helena',
-  },
-  {
-    class: 'filter-hudson',
-    name: 'Hudson',
-  },
-  {
-    class: 'filter-inkwell',
-    name: 'Inkwell',
-  },
-  {
-    class: 'filter-kelvin',
-    name: 'Kelvin',
-  },
-  {
-    class: 'filter-juno',
-    name: 'Kuno',
-  },
-  {
-    class: 'filter-lark',
-    name: 'Lark',
-  },
-  {
-    class: 'filter-lofi',
-    name: 'Lo-Fi',
-  },
-  {
-    class: 'filter-ludwig',
-    name: 'Ludwig',
-  },
-  {
-    class: 'filter-maven',
-    name: 'Maven',
-  },
-  {
-    class: 'filter-mayfair',
-    name: 'Mayfair',
-  },
-  {
-    class: 'filter-moon',
-    name: 'Moon',
-  },
-  {
-    class: 'filter-nashville',
-    name: 'Nashville',
-  },
-  {
-    class: 'filter-perpetua',
-    name: 'Perpetua',
-  },
-  {
-    class: 'filter-poprocket',
-    name: 'Poprocket',
-  },
-  {
-    class: 'filter-reyes',
-    name: 'Reyes',
-  },
-  {
-    class: 'filter-rise',
-    name: 'Rise',
-  },
-  {
-    class: 'filter-sierra',
-    name: 'Sierra',
-  },
-  {
-    class: 'filter-skyline',
-    name: 'Skyline',
-  },
-  {
-    class: 'filter-slumber',
-    name: 'Slumber',
-  },
-  {
-    class: 'filter-stinson',
-    name: 'Stinson',
-  },
-  {
-    class: 'filter-sutro',
-    name: 'Sutro',
-  },
-  {
-    class: 'filter-toaster',
-    name: 'Toaster',
-  },
-  {
-    class: 'filter-valencia',
-    name: 'Valencia',
-  },
-  {
-    class: 'filter-vesper',
-    name: 'Vesper',
-  },
-  {
-    class: 'filter-walden',
-    name: 'Walden',
-  },
-  {
-    class: 'filter-willow',
-    name: 'Willow',
-  },
-  {
-    class: 'filter-xpro-ii',
-    name: 'X-Pro II',
+    name: 'sharpen',
   },
 ]
